added review/edit. Added star rating radiobuttons

+++ movievultures/src/main/java/movievultures/model/dao/ReviewDao.java
@@ -7,7 +7,7 @@ import movievultures.model.Review;
 import movievultures.model.User;
 
 public interface ReviewDao {
-	Review getReview( Long id );
+	Review getReview( int id );
 	List<Review> getReviewsByUser( User user );
 	List<Review> getReviewsByMovie( Movie movie );
 	Review saveReview( Review review );
